// src/types.ts
export type NodeState = {
  killed: boolean; // this is used to know if the node was stopped by the /stop route. It's important for the unit tests but not very relevant for the Ben-Or implementation
  x: Value | null; // the current consensus value
  decided: boolean | null; // used to know if the node reached finality
  k: number | null; // current step of the node
};

export type Value = 0 | 1 | "?";

export type Message = {
  type: string;
  k: number;
  x: Value;
};

export type MessageCount = {
  n: number;
  nb0: number;
  nb1: number;
};